sqlite_to_postgres/requirements.txt
isort==5.10.1
psycopg2-binary==2.9.2
python-environ==0.4.54
